Tweaked tests
