+++ src/test/groovy/graphql/execution/instrumentation/dataloader/models/Shop.java
@@ -1,16 +1,16 @@
-package graphql.execution.instrumentation.dataloader;
+package graphql.execution.instrumentation.dataloader.models;
 
 import java.util.List;
 
-public class Department {
+public class Shop {
     private final String id;
     private final String name;
-    private final List<String> productIds;
+    private final List<String> departmentIds;
 
-    public Department(String id, String name, List<String> productIds) {
+    public Shop(String id, String name, List<String> departmentIds) {
         this.id = id;
         this.name = name;
-        this.productIds = productIds;
+        this.departmentIds = departmentIds;
     }
 
     public String getId() {
@@ -21,16 +21,16 @@ public class Department {
         return name;
     }
 
-    public List<String> getProductIds() {
-        return productIds;
+    public List<String> getDepartmentIds() {
+        return departmentIds;
     }
 
     @Override
     public String toString() {
-        return "Department{" +
+        return "Shop{" +
                 "id='" + id + '\'' +
                 ", name='" + name + '\'' +
-                ", productIds=" + productIds +
+                ", departmentIds=" + departmentIds +
                 '}';
     }
 }
